Silence compiler warnings for the x64 build.

+++ trunk/StExBar/src/Filter.cpp
@@ -1,6 +1,6 @@
 // StExBar - an explorer toolbar
 
-// Copyright (C) 2007-2009 - Stefan Kueng
+// Copyright (C) 2007-2010 - Stefan Kueng
 
 // This program is free software; you can redistribute it and/or
 // modify it under the terms of the GNU General Public License
@@ -157,7 +157,7 @@ bool CDeskBand::Filter(LPTSTR filter)
 										{
 											m_noShows.erase(m_noShows.begin() + i);
 											i--;
-											UINT puItem = i;
+											UINT puItem = (UINT)i;
 											pShellFolderView->AddObject(pidlNoShow, &puItem);
 											CoTaskMemFree(pidlNoShow);
 										}
